requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astunparse==1.6.3
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
<<<<<<< HEAD
gunicorn==20.0.4
=======
>>>>>>> 099d1aa63d6464a2c1400ab0d11bbe706a266c14
h5py==2.10.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
lockfile==0.12.2
Markdown==3.2.2
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.19.0
oauthlib==3.1.0
<<<<<<< HEAD
=======
opencv-python==4.3.0.36
>>>>>>> 099d1aa63d6464a2c1400ab0d11bbe706a266c14
opt-einsum==3.2.1
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
<<<<<<< HEAD
pkg-resources==0.0.0
=======
>>>>>>> 099d1aa63d6464a2c1400ab0d11bbe706a266c14
progress==1.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
pytoml==0.1.21
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scipy==1.4.1
six==1.14.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
